moving getUserMedia - created a new class CameraViewRenderer
